+++ spec/fixtures/cookbooks/oneview_test/recipes/logical_switch_refresh.rb
@@ -1,6 +1,6 @@
 #
 # Cookbook Name:: oneview_test
-# Recipe:: storage_pool_add_with_ip
+# Recipe:: logical_switch_reset_connection_template
 #
 # (c) Copyright 2016 Hewlett Packard Enterprise Development LP
 #
@@ -14,7 +14,7 @@
 # specific language governing permissions and limitations under the License.
 #
 
-oneview_storage_pool 'StoragePool1' do
+oneview_logical_switch 'LogicalSwitch4' do
   client node['oneview_test']['client']
-  storage_system_ip '10.1.1.1'
+  action :refresh
 end