+++ spec/unit/resources/matchers_spec.rb
@@ -51,8 +51,10 @@ describe 'oneview_test::default' do
     expect(chef_run).to_not create_oneview_fcoe_network_if_missing('')
     expect(chef_run).to_not delete_oneview_fcoe_network('')
 
-    # oneview_firmware_bundle
-    expect(chef_run).to_not add_oneview_firmware_bundle('')
+    # oneview_firmware
+    expect(chef_run).to_not add_oneview_firmware('')
+    expect(chef_run).to_not remove_oneview_firmware('')
+    expect(chef_run).to_not create_oneview_firmware_custom_spp('')
 
     # oneview_interconnect
     expect(chef_run).to_not set_oneview_interconnect_uid_light('')
@@ -76,10 +78,22 @@ describe 'oneview_test::default' do
     expect(chef_run).to_not create_oneview_logical_switch_group_if_missing('')
     expect(chef_run).to_not delete_oneview_logical_switch_group('')
 
+    # oneview_logical_switch
+    expect(chef_run).to_not create_oneview_logical_switch('')
+    expect(chef_run).to_not create_oneview_logical_switch_if_missing('')
+    expect(chef_run).to_not delete_oneview_logical_switch('')
+    expect(chef_run).to_not refresh_oneview_logical_switch('')
+
     # oneview_network_set
     expect(chef_run).to_not create_oneview_network_set('')
     expect(chef_run).to_not create_oneview_network_set_if_missing('')
     expect(chef_run).to_not delete_oneview_network_set('')
+
+    # oneview_power_device
+    expect(chef_run).to_not add_oneview_power_device('')
+    expect(chef_run).to_not add_oneview_power_device_if_missing('')
+    expect(chef_run).to_not discover_oneview_power_device('')
+    expect(chef_run).to_not remove_oneview_power_device('')
 
     # oneview_rack
     expect(chef_run).to_not add_oneview_rack('')
@@ -99,12 +113,18 @@ describe 'oneview_test::default' do
     expect(chef_run).to_not remove_oneview_server_hardware_type('')
 
     # oneview_storage_pool
-    expect(chef_run).to_not add_oneview_storage_pool('')
+    expect(chef_run).to_not add_oneview_storage_pool_if_missing('')
     expect(chef_run).to_not remove_oneview_storage_pool('')
 
     # oneview_storage_system
     expect(chef_run).to_not add_oneview_storage_system('')
+    expect(chef_run).to_not add_oneview_storage_system_if_missing('')
+    expect(chef_run).to_not edit_oneview_storage_system_credentials('')
     expect(chef_run).to_not remove_oneview_storage_system('')
+
+    # oneview_switch
+    expect(chef_run).to_not remove_oneview_switch('')
+    expect(chef_run).to_not none_oneview_switch('')
 
     # oneview_volume
     expect(chef_run).to_not create_oneview_volume('')