+++ spec/unit/resources/server_hardware_type/edit_spec.rb
@@ -4,10 +4,9 @@ describe 'oneview_test::server_hardware_type_edit' do
   let(:resource_name) { 'server_hardware_type' }
   include_context 'chef context'
 
-  it 'does nothing when it does not exists' do
+  it 'raise error when it does not exists' do
     allow_any_instance_of(OneviewSDK::ServerHardwareType).to receive(:exists?).and_return(false)
-    expect_any_instance_of(OneviewSDK::ServerHardwareType).to_not receive(:update)
-    expect(real_chef_run).to edit_oneview_server_hardware_type('ServerHardwareType1')
+    expect { real_chef_run }.to raise_error(/ServerHardwareType1 does not exists/)
   end
 
   it 'updates it when it exists but not alike' do